CODE/web_server/server/requirements.txt
Django>=3.0,<4.0
djangorestframework==3.12.4
pandas==1.3.3
scikit_maad==1.2
tensorflow==2.7.0
wavfile==2.2.0
scipy==1.7.2
keras==2.7.0
librosa==0.8.1
numpy==1.19.5
matplotlib
